// src/sdks/Newcomputation/computationV2ApiDocs.ts
export interface EnumResp {
  code?: number;
  name?: string;
  score?: string;
  subList?: EnumResp[];
}