// src/sdks/systemV2ApiDocs.ts
export interface DictTypeResp {
  dictName?: string;

  dictType?: string;

  id?: number;
}
// ...
export interface OrgPojo {
  id?: number;

  orgCode?: string;

  orgName?: string;

  orgType?: '0' | '1' | '2' | '3';

  pid?: number;

  orgStatus?: number;
}
// ...
export interface Tree {
  children?: Tree[];
  code?: number;
  name?: string;
  pcode?: number;
}